+++ src/main/java/com/example/eightyage/domain/auth/tokenstate/TokenState.java
@@ -1,4 +1,4 @@
-package com.example.eightyage.domain.auth.entity;
+package com.example.eightyage.domain.auth.tokenstate;
 
 public enum TokenState {
     VALID,
