+++ src/main/java/com/example/eightyage/global/exception/BadRequestException.java
@@ -1,6 +1,6 @@
 package com.example.eightyage.global.exception;
 
-public class BadRequestException extends CustomException {
+public class BadRequestException extends HandledException {
 
     public BadRequestException() {
         super(ErrorCode.BAD_REQUEST);
