+++ src/main/java/com/example/eightyage/global/exception/HandledException.java
@@ -4,20 +4,20 @@ import lombok.Getter;
 import org.springframework.http.HttpStatus;
 
 @Getter
-public class CustomException extends RuntimeException {
+public class HandledException extends RuntimeException {
 
     private final HttpStatus httpStatus;
     private final String message;
 
     // 예외 던질시 기본 메세지 출력
-    public CustomException(ErrorCode errorCode) {
+    public HandledException(ErrorCode errorCode) {
         super(errorCode.getDefaultMessage());
         this.httpStatus = errorCode.getStatus();
         this.message = errorCode.getDefaultMessage();
     }
 
     // 예외 던질시 메세지 출력
-    public CustomException(ErrorCode errorCode, String message) {
+    public HandledException(ErrorCode errorCode, String message) {
         super(message);
         this.httpStatus = errorCode.getStatus();
         this.message = message;
