+++ src/main/java/com/example/eightyage/global/exception/UnauthorizedException.java
@@ -1,6 +1,6 @@
 package com.example.eightyage.global.exception;
 
-public class UnauthorizedException extends CustomException {
+public class UnauthorizedException extends HandledException {
 
   public UnauthorizedException() {
     super(ErrorCode.AUTHORIZATION);
